# crates/nexq-api-rest/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NexQ",
    "description": "NexQ's native API: the complete operation set, including the extensions the SQS-compatible facade cannot express. This document is generated from the server's own routing table and is the source every client is generated from.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/queues": {
      "get": {
        "summary": "List queues",
        "description": "In name order, one page at a time, optionally filtered by `prefix`.\n\nWhen `next_cursor` is not null there are more: pass it back as `cursor` to continue. It is a **cursor, not an offset** — it names where to resume, so a queue created or deleted between two requests cannot make you skip one or see it twice. That is the storage layer's keyset guarantee carried onto the wire rather than re-derived from it, and it is why there is no \"page number\".\n\nA cursor is this server's to issue. Constructing one yourself is not supported and is refused rather than interpreted. An empty result is a normal answer, not an error.",
        "operationId": "listQueues",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "description": "Resume after this queue — the `next_cursor` from the previous page.",
            "schema": {
              "description": "Resume after this queue — the `next_cursor` from the previous page.",
              "type": [
                "string",
                "null"
              ],
              "default": null
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "limit",
            "description": "How many to return at most.",
            "schema": {
              "description": "How many to return at most.",
              "type": [
                "integer",
                "null"
              ],
              "format": "uint",
              "default": null,
              "maximum": 1000,
              "minimum": 1
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "prefix",
            "description": "Only queues whose name starts with this.",
            "schema": {
              "description": "Only queues whose name starts with this.",
              "type": [
                "string",
                "null"
              ],
              "default": null
            },
            "style": "form"
          }
        ],
        "responses": {
          "default": {
            "description": "The body of every failed request.\n\n Nested under `error` rather than flattened so that a successful response and a failed\n one can never be told apart only by which fields happen to be present.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "200": {
            "description": "One page of queues, with a cursor when more remain. `queues` is an empty list rather than absent when there are none.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListQueuesResponse"
                }
              }
            }
          },
          "400": {
            "description": "The limit is outside 1–1000, a query parameter is not one this endpoint takes, or the cursor was not one this server issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "The bearer token is missing or does not check out.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The token is `<key_id>.<secret>` for a credential in the server's registry. Presented in full on every request, so it should travel over TLS."
      }
    },
    "schemas": {
      "ErrorBody": {
        "description": "The body of every failed request.\n\n Nested under `error` rather than flattened so that a successful response and a failed\n one can never be told apart only by which fields happen to be present.",
        "type": "object",
        "properties": {
          "error": {
            "description": "What went wrong.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              }
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "ErrorDetail": {
        "description": "The failure itself.",
        "type": "object",
        "properties": {
          "code": {
            "description": "A stable, machine-readable identifier for this kind of failure. Safe to branch on.",
            "type": "string"
          },
          "message": {
            "description": "A human-readable explanation. May change between releases; do not parse it.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "ListQueuesResponse": {
        "description": "One page of queues.",
        "type": "object",
        "properties": {
          "next_cursor": {
            "description": "Pass as `cursor` to fetch the next page, or `null` when this is the last one.\n\n A **cursor**, not an offset: it names where to resume, so a queue created or\n deleted between pages cannot make a caller skip one or see it twice. Present\n whenever more remain, including when `limit` was not given.",
            "type": [
              "string",
              "null"
            ]
          },
          "queues": {
            "description": "This page, in name order. Empty rather than absent when there are none.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueueResponse"
            }
          }
        },
        "required": [
          "queues"
        ]
      },
...
      "QueueAttributesResponse": {
        "description": "A queue's settings.",
        "type": "object",
        "properties": {
          "delay_seconds": {
            "description": "How long a newly sent message waits before it can be received.",
            "type": "integer",
            "format": "uint64",
            "minimum": 0
          },
          "receive_wait_time_seconds": {
            "description": "How long a receive that finds nothing waits by default — the long poll a consumer\n gets without asking for one.",
            "type": "integer",
            "format": "uint64",
            "minimum": 0
          },
          "visibility_timeout_seconds": {
            "description": "How long a claimed message stays invisible to other consumers before being\n redelivered.",
            "type": "integer",
            "format": "uint64",
            "minimum": 0
          }
        },
        "required": [
          "visibility_timeout_seconds",
          "delay_seconds",
          "receive_wait_time_seconds"
        ]
      },
...
      "QueueResponse": {
        "description": "A queue, as this API represents it.",
        "type": "object",
        "properties": {
          "attributes": {
            "description": "How the queue is configured.",
            "allOf": [
              {
                "$ref": "#/components/schemas/QueueAttributesResponse"
              }
            ]
          },
          "created_at": {
            "description": "When the queue was created.",
            "type": "string",
            "format": "date-time"
          },
          "last_modified_at": {
            "description": "When its attributes were last changed, or when it was created if they never have\n been. Reported separately because answering with the creation time for both would\n be a plausible-looking lie.",
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "description": "The queue's name, which is also its address: `/api/v1/queues/{name}`.",
            "type": "string"
          }
        },
        "required": [
          "name",
          "created_at",
          "last_modified_at",
          "attributes"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}